requirements.txt
gymnasium>=0.29
highway-env>=1.8.2

stable-baselines3>=2.3.0

torch
numpy>=1.23
matplotlib>=3.7
moviepy>=1.0.3
pandas>=2.0
tqdm>=4.66
pillow>=10.0
